+++ src/adler32.c
@@ -33,16 +33,16 @@
 #define A32_BASE 65521
 #define A32_NMAX 5552
 
-unsigned long tinf_adler32(const void *data, unsigned long length)
+uint32_t tinf_adler32(const void *data, uint32_t length)
 {
     const unsigned char *buf = (const unsigned char *) data;
 
-    unsigned long s1 = 1;
-    unsigned long s2 = 0;
+    uint32_t s1 = 1;
+    uint32_t s2 = 0;
 
     while (length > 0) {
-        long k = length < A32_NMAX ? length : A32_NMAX;
-        long i;
+        uint32_t k = length < A32_NMAX ? length : A32_NMAX;
+        uint32_t i;
 
         for (i = k / 16; i; --i, buf += 16) {
             s1 += buf[0];
